+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+* Added:
+  * There is now a "synchronize" toggle (link icon in the top bar),
+    which synchronizes play/pause/seek events for all media of the same category
+    (image, audio, video).
 * Fixed:
   * When a source was specified on the command line,
     you had to click "add player" for it to start playing.